rename pydepsi to depsi
