+++ README.md
@@ -1,6 +1,5 @@
-# PyDePSI (Tentative name)
-
-This is repository is WIP, where we are developing a Python package for inteferometric SAR processing. The software will be inspired by the MATLAB software DePSI, but implemented in Python and include recent developments in the field.
+# DePSI 
+This is the WIP repository for the Python version of DePSI *(), a Python package for inteferometric SAR processing. The software is inspired by the MATLAB version DePSI. In this repository, we implement classic DePSI algorithms and new InSAR developments in Python.
 
 ## Installation for development
 
@@ -9,14 +8,14 @@ It is assumed that you have `mamba` installed. If not, you can find the installa
 Clone this repository and `cd` into it:
 
 ```bash
-git clone git@github.com:MotionbyLearning/PyDePSI.git
-cd PyDePSI
+git clone git@github.com:TUDelftGeodesy/DePSI.git
+cd DePSI
 ```
 
-Create a new conda environment (here we give an example name `pydepsi-dev`) with `mamba`.:
+Create a new conda environment (here we give an example name `depsi-dev`) with `mamba`.:
 
 ```bash
-mamba create -c conda-forge -n pydepsi-dev python=3.12
+mamba create -c conda-forge -n depsi-dev python=3.12
 ```
 
 Here we use Python 3.12 since we aim to support python 3.10 and above.
@@ -24,7 +23,7 @@ Here we use Python 3.12 since we aim to support python 3.10 and above.
 Activate the environment:
 
 ```bash
-mamba activate pydepsi-dev
+mamba activate depsi-dev
 ```
 
 Install this package in development mode:
