+++ examples/scripts/script_reslc.py
@@ -15,8 +15,8 @@ from dask.distributed import Client
 from dask_jobqueue import SLURMCluster
 import sarxarray
 
-from pydepsi.slc import ifg_to_slc
-from pydepsi.io import read_metadata
+from depsi.slc import ifg_to_slc
+from depsi.io import read_metadata
 
 # Make a logger to log the stages of processing
 logger = logging.getLogger(__name__)
