+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "pydepsi"
+name = "depsi"
 version = "0.1.0"
 requires-python = ">=3.10"
 dependencies = [
@@ -48,7 +48,7 @@ demo = [
 ]
 
 [tool.setuptools]
-packages = ["pydepsi"]
+packages = ["depsi"]
 
 [tool.pytest.ini_options]
 minversion = "6.0"
@@ -59,7 +59,7 @@ testpaths = [
 
 [tool.coverage.run]
 branch = true
-source = ["pydepsi"]
+source = ["depsi"]
 
 [tool.ruff]
 lint.select = [
